Add `colour.utilities.requirements` module.
