+++ colour/utilities/__init__.py
@@ -4,6 +4,42 @@ import sys
 
 from colour.hints import Any
 
+from .verbose import (
+    MixinLogging,
+    ColourWarning,
+    ColourUsageWarning,
+    ColourRuntimeWarning,
+    message_box,
+    show_warning,
+    warning,
+    runtime_warning,
+    usage_warning,
+    filter_warnings,
+    as_bool,
+    suppress_warnings,
+    suppress_stdout,
+    numpy_print_options,
+    ANCILLARY_COLOUR_SCIENCE_PACKAGES,
+    ANCILLARY_RUNTIME_PACKAGES,
+    ANCILLARY_DEVELOPMENT_PACKAGES,
+    ANCILLARY_EXTRAS_PACKAGES,
+    describe_environment,
+    multiline_str,
+    multiline_repr,
+)
+from .requirements import (
+    is_ctlrender_installed,
+    is_graphviz_installed,
+    is_matplotlib_installed,
+    is_networkx_installed,
+    is_opencolorio_installed,
+    is_openimageio_installed,
+    is_pandas_installed,
+    is_tqdm_installed,
+    is_trimesh_installed,
+    is_xxhash_installed,
+    required,
+)
 from .data_structures import (
     Lookup,
     Structure,
@@ -31,18 +67,6 @@ from .common import (
     batch,
     disable_multiprocessing,
     multiprocessing_pool,
-    is_ctlrender_installed,
-    is_graphviz_installed,
-    is_matplotlib_installed,
-    is_networkx_installed,
-    is_opencolorio_installed,
-    is_openimageio_installed,
-    is_pandas_installed,
-    is_tqdm_installed,
-    is_trimesh_installed,
-    is_xxhash_installed,
-    required,
-    as_bool,
     is_iterable,
     is_numeric,
     is_integer,
@@ -55,27 +79,6 @@ from .common import (
     optional,
     slugify,
     int_digest,
-)
-from .verbose import (
-    ColourWarning,
-    ColourUsageWarning,
-    ColourRuntimeWarning,
-    message_box,
-    show_warning,
-    warning,
-    runtime_warning,
-    usage_warning,
-    filter_warnings,
-    suppress_warnings,
-    suppress_stdout,
-    numpy_print_options,
-    ANCILLARY_COLOUR_SCIENCE_PACKAGES,
-    ANCILLARY_RUNTIME_PACKAGES,
-    ANCILLARY_DEVELOPMENT_PACKAGES,
-    ANCILLARY_EXTRAS_PACKAGES,
-    describe_environment,
-    multiline_str,
-    multiline_repr,
 )
 from .array import (
     MixinDataclassFields,
@@ -133,6 +136,42 @@ from colour.utilities.deprecation import ModuleAPI, build_API_changes
 from colour.utilities.documentation import is_documentation_building
 
 __all__ = [
+    "MixinLogging",
+    "ColourWarning",
+    "ColourUsageWarning",
+    "ColourRuntimeWarning",
+    "message_box",
+    "show_warning",
+    "warning",
+    "runtime_warning",
+    "usage_warning",
+    "filter_warnings",
+    "as_bool",
+    "suppress_warnings",
+    "suppress_stdout",
+    "numpy_print_options",
+    "ANCILLARY_COLOUR_SCIENCE_PACKAGES",
+    "ANCILLARY_RUNTIME_PACKAGES",
+    "ANCILLARY_DEVELOPMENT_PACKAGES",
+    "ANCILLARY_EXTRAS_PACKAGES",
+    "describe_environment",
+    "multiline_str",
+    "multiline_repr",
+]
+__all__ += [
+    "is_ctlrender_installed",
+    "is_graphviz_installed",
+    "is_matplotlib_installed",
+    "is_networkx_installed",
+    "is_opencolorio_installed",
+    "is_openimageio_installed",
+    "is_pandas_installed",
+    "is_tqdm_installed",
+    "is_trimesh_installed",
+    "is_xxhash_installed",
+    "required",
+]
+__all__ += [
     "Lookup",
     "Structure",
     "CanonicalMapping",
@@ -159,18 +198,6 @@ __all__ += [
     "batch",
     "disable_multiprocessing",
     "multiprocessing_pool",
-    "is_ctlrender_installed",
-    "is_graphviz_installed",
-    "is_matplotlib_installed",
-    "is_networkx_installed",
-    "is_opencolorio_installed",
-    "is_openimageio_installed",
-    "is_pandas_installed",
-    "is_tqdm_installed",
-    "is_trimesh_installed",
-    "is_xxhash_installed",
-    "required",
-    "as_bool",
     "is_iterable",
     "is_numeric",
     "is_integer",
@@ -183,27 +210,6 @@ __all__ += [
     "optional",
     "slugify",
     "int_digest",
-]
-__all__ += [
-    "ColourWarning",
-    "ColourUsageWarning",
-    "ColourRuntimeWarning",
-    "message_box",
-    "show_warning",
-    "warning",
-    "runtime_warning",
-    "usage_warning",
-    "filter_warnings",
-    "suppress_warnings",
-    "suppress_stdout",
-    "numpy_print_options",
-    "ANCILLARY_COLOUR_SCIENCE_PACKAGES",
-    "ANCILLARY_RUNTIME_PACKAGES",
-    "ANCILLARY_DEVELOPMENT_PACKAGES",
-    "ANCILLARY_EXTRAS_PACKAGES",
-    "describe_environment",
-    "multiline_str",
-    "multiline_repr",
 ]
 __all__ += [
     "MixinDataclassFields",
